# docs/docs-site/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Community Feed API",
    "description": "Community-governed Bluesky feed generator where subscribers democratically vote on ranking algorithm weights. Built on AT Protocol.\n\n## Authentication\n- **Governance endpoints** require a session cookie or bearer token from `POST /api/governance/auth/login`.\n- **Admin endpoints** additionally require the caller's DID to be in the `BOT_ADMIN_DIDS` allowlist.\n- **Feed endpoints** are public (called by the Bluesky app). Auth is optional for subscriber tracking.\n- **Transparency endpoints** are public and unauthenticated.",
    "version": "1.1.0",
    "contact": {
      "name": "Community Feed",
      "url": "https://github.com/andrewnordstrom-eng/bluesky-community-feed"
    },
    "license": {
      "name": "MIT"
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "governance_session",
        "description": "Session cookie from POST /api/governance/auth/login"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token from POST /api/governance/auth/login (for CLI/MCP)"
      }
    },
...
  },
  "paths": {
...
    "/api/transparency/audit": {
      "get": {
        "summary": "Governance audit log",
        "tags": [
          "Transparency"
        ],
        "description": "Returns a paginated, append-only audit log of all governance actions. Vote details are redacted to preserve voter privacy (shows counts, not specific votes).",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "entries": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "action": {
                            "type": "string",
                            "description": "Action type (e.g. vote_cast, epoch_created)"
                          },
                          "actor_did": {
                            "type": "string",
                            "nullable": true,
                            "description": "Redacted for privacy"
                          },
                          "epoch_id": {
                            "type": "integer",
                            "nullable": true
                          },
                          "details": {
                            "type": "object",
                            "additionalProperties": true,
                            "description": "Action-specific details (vote details redacted)"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "offset": {
                          "type": "integer"
                        },
                        "has_more": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "total",
                        "limit",
                        "offset",
                        "has_more"
                      ]
                    }
                  },
                  "required": [
                    "entries",
                    "pagination"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error code (e.g. \"ValidationError\", \"NotFound\")"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Request correlation ID for debugging"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error code (e.g. \"ValidationError\", \"NotFound\")"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Request correlation ID for debugging"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://feed.yourdomain.com",
      "description": "Production"
    }
  ],
...
}